Create FirebaseLists component
